@@ -7,7 +7,7 @@ import typing as t
 import yaml
 
 if t.TYPE_CHECKING:
-    from typing_extensions import Literal, TypeAlias
+    from typing import Literal, TypeAlias
 
     FormatLiteral = Literal["json", "yaml"]
 
